Fix transaction decoder

+++ pkg/transactionparser/transaction/abstract.go
@@ -6,7 +6,7 @@ import (
 
 // DecodedTransactioner is an interface for transaction decoded from raw bytes.
 type DecodedTransactioner interface {
-	GetType() int
+	GetType() uint8
 	GetAmount() uint64
 	GetCounter() uint64
 	GetReceiver() address.Address
